# openapi.yaml
openapi: 3.0.3
info:
  title: PredictMarket SaaS API
  version: 0.1.0
  description: Merchant prediction-market API backed by PostgreSQL, Redis, and NATS JetStream.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/merchants/register:
    post:
      tags: [Merchants]
      summary: Register merchant
      security: []
      requestBody: {$ref: "#/components/requestBodies/MerchantRegistration"}
      responses:
        "201": {$ref: "#/components/responses/Created"}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/merchants:
    get:
      tags: [Merchants]
      summary: List merchants
      security: [{adminBearer: []}]
      parameters: [{$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/Limit"}]
      responses: {"200": {$ref: "#/components/responses/OK"}, "401": {$ref: "#/components/responses/Error"}}
  /api/v1/merchants/{merchantID}/config:
    parameters: [{$ref: "#/components/parameters/MerchantID"}]
    get:
      tags: [Merchants]
      summary: Get merchant configuration
      responses: {"200": {$ref: "#/components/responses/OK"}, "403": {$ref: "#/components/responses/Error"}}
    patch:
      tags: [Merchants]
      summary: Update merchant configuration
      requestBody: {$ref: "#/components/requestBodies/MerchantUpdate"}
      responses: {"200": {$ref: "#/components/responses/OK"}, "400": {$ref: "#/components/responses/Error"}}
  /api/v1/merchants/{merchantID}/v3-secret/reissue:
    post:
      tags: [Merchants]
      summary: Reissue a merchant V3 HMAC secret
      security: [{adminBearer: []}]
      parameters: [{$ref: "#/components/parameters/MerchantID"}]
      responses: {"201": {$ref: "#/components/responses/Created"}, "401": {$ref: "#/components/responses/Error"}, "404": {$ref: "#/components/responses/Error"}, "503": {$ref: "#/components/responses/Error"}}
...
  /api/v1/admin/callback-dead-letters/{channel}/{outboxID}/replay:
    post:
      tags: [V2 Integration]
      summary: Replay one dead-lettered callback or webhook
      security: [{adminBearer: []}]
      parameters:
        - {name: channel, in: path, required: true, schema: {type: string, enum: [callback, webhook]}}
        - {name: outboxID, in: path, required: true, schema: {type: string, format: uuid}}
      responses: {"204": {$ref: "#/components/responses/NoContent"}, "401": {$ref: "#/components/responses/Error"}, "404": {$ref: "#/components/responses/Error"}}
  /api/v1/merchants/{merchantID}/integration/verify-callback:
    post:
      tags: [V2 Integration]
      summary: Prove callback URL ownership with a signed challenge echo
      security: [{adminBearer: []}]
      parameters: [{name: merchantID, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {"200": {$ref: "#/components/responses/OK"}, "401": {$ref: "#/components/responses/Error"}, "404": {$ref: "#/components/responses/Error"}, "502": {$ref: "#/components/responses/Error"}}
  /api/v1/merchants/{merchantID}/integration/reset-degraded:
    post:
      tags: [V2 Integration]
      summary: Clear a merchant's seamless degraded state
      security: [{adminBearer: []}]
      parameters: [{name: merchantID, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {"200": {$ref: "#/components/responses/OK"}, "401": {$ref: "#/components/responses/Error"}}
  /api/v1/admin/markets/{marketID}/void:
    post:
      tags: [Admin]
      summary: Void an unsettled market and refund every order in full
      security: [{adminBearer: []}]
      parameters: [{name: marketID, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {"200": {$ref: "#/components/responses/OK"}, "401": {$ref: "#/components/responses/Error"}, "404": {$ref: "#/components/responses/Error"}, "409": {$ref: "#/components/responses/Error"}}
...
  /api/v1/wallets/{userID}:
    get:
      tags: [Wallets]
      summary: Get wallet balance
      parameters:
        - {$ref: "#/components/parameters/UserID"}
        - {name: currency, in: query, schema: {$ref: "#/components/schemas/Currency"}}
      responses: {"200": {$ref: "#/components/responses/OK"}, "404": {$ref: "#/components/responses/Error"}}
  /api/v1/wallets/{userID}/credit:
    post:
      tags: [Wallets]
      summary: Credit virtual wallet
      parameters: [{$ref: "#/components/parameters/UserID"}, {$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {$ref: "#/components/requestBodies/WalletCredit"}
      responses: {"200": {$ref: "#/components/responses/OK"}, "428": {$ref: "#/components/responses/Error"}}
  /api/v1/wallets/{userID}/transactions:
    get:
      tags: [Wallets]
      summary: List transactions
      parameters:
        - {$ref: "#/components/parameters/UserID"}
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/Limit"}
      responses: {"200": {$ref: "#/components/responses/OK"}}
...
components:
...
  securitySchemes:
    merchantBearer: {type: http, scheme: bearer, bearerFormat: API key}
    adminBearer: {type: http, scheme: bearer, bearerFormat: Admin API key}
    v2SignedMerchant: {type: http, scheme: bearer, bearerFormat: API key plus X-PM-Signature and X-PM-Timestamp HMAC headers}
    userSessionBearer: {type: http, scheme: bearer, bearerFormat: Hosted browser session JWT}
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    MerchantID: {name: merchantID, in: path, required: true, schema: {type: string, format: uuid}}
...
    UserID: {name: userID, in: path, required: true, schema: {type: string, maxLength: 255}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, description: Unique key used to make POST retries safe, schema: {type: string, minLength: 1, maxLength: 255}}
...
  responses:
    OK: {description: Success, content: {application/json: {schema: {$ref: "#/components/schemas/DataEnvelope"}}}}
    Created: {description: Created, content: {application/json: {schema: {$ref: "#/components/schemas/DataEnvelope"}}}}
    NoContent: {description: Operation completed}
    Error: {description: Error, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}
  requestBodies:
    MerchantRegistration:
      required: true
      content: {application/json: {schema: {$ref: "#/components/schemas/MerchantRegistration"}}}
    MerchantUpdate:
      required: true
      content: {application/json: {schema: {$ref: "#/components/schemas/MerchantUpdate"}}}
...
    WalletCredit:
      required: true
      content: {application/json: {schema: {$ref: "#/components/schemas/WalletCredit"}}}
...
  schemas:
    DataEnvelope: {type: object, required: [data], properties: {data: {}, meta: {type: object, additionalProperties: true}}}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
    Currency: {type: string, enum: [USD, EUR, CNY, JPY, GBP]}
...
    MerchantRegistration:
      type: object
      required: [name, email, currency, timezone]
      additionalProperties: false
      properties:
        name: {type: string, maxLength: 255}
        email: {type: string, format: email, maxLength: 320}
        currency: {$ref: "#/components/schemas/Currency"}
        timezone: {type: string, example: Asia/Shanghai}
    MerchantUpdate:
      type: object
      additionalProperties: false
      properties:
        name: {type: string}
        currency: {$ref: "#/components/schemas/Currency"}
        timezone: {type: string}
...
    WalletCredit:
      type: object
      required: [currency, amount]
      additionalProperties: false
      properties:
        currency: {$ref: "#/components/schemas/Currency"}
        amount: {type: number, minimum: 0, exclusiveMinimum: true}
        type: {type: string, enum: [credit, admin_credit]}
...
security:
  - merchantBearer: []